+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Raise an exception when a CAN transceiver's driver is not installed.
 - Add optional password for the FOE bootloader.
 - Add strict mode to the load configuration method.
+- Store nack error code on NACK exceptions
 
 ### Fixed
 - SDO Error after a store/restore parameters operation.